requirements.txt
appdirs==1.4.4
asgiref==3.2.10
backcall==0.2.0
black==20.8b1
click==7.1.2
decorator==4.4.2
Django==3.0.5
djangorestframework==3.11.1
djongo==1.3.3
ipdb==0.13.3
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
mypy-extensions==0.4.3
mysqlclient==2.0.1
parso==0.7.1
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==3.0.7
ptyprocess==0.6.0
Pygments==2.7.1
pymongo==3.11.0
pytz==2020.1
regex==2020.7.14
sqlparse==0.2.4
toml==0.10.1
traitlets==5.0.4
typed-ast==1.4.1
typing-extensions==3.7.4.3
wcwidth==0.2.5
